esp_ai_engine/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-constraint==1.4.0
google-generativeai==0.3.1
requests==2.31.0
Werkzeug==2.3.7
protobuf==4.24.4
